xlnx: rename 'assists/xlnx_pm/*' to 'assists/xlnx/*'
Use "xlnx" subdir under assists for organizing the plugin helpers for
generation of xilinx specific subsystems and protections cdo
configurations.
